Added resume functionality to download

+++ examples/pyggy/download-pyg.sh
@@ -16,21 +16,26 @@ if [ -f "models/$FILENAME" ]; then
     echo -e "${YELLOW}File already exists and MD5 hash matches!${NC}"
     exit 0
   else
-    echo -e "${YELLOW}File already exists but MD5 hash doesn't match. Deleting file and re-downloading.${NC}"
-    rm -f models/$FILENAME
+    read -p "$(echo -e "${YELLOW}File already exists but MD5 hash doesn't match. Do you want to resume the download? [y/n]${NC}")" -n 1 -r
+    echo
+    if [[ $REPLY =~ ^[Yy]$ ]]; then
+      echo -e "${GREEN}Resuming download...${NC}"
+      wget -c -q --show-progress $URL -P models/ -O $FILENAME
+    else
+      echo -e "${YELLOW}Deleting file and re-downloading.${NC}"
+      rm -f models/$FILENAME
+    fi
   fi
 fi
 
-echo -e "${GREEN}Downloading file...${NC}"
-wget -q --show-progress $URL -P models/ -O $FILENAME
+if [ ! -f "models/$FILENAME" ]; then
+  echo -e "${GREEN}Downloading file...${NC}"
+  wget -c -q --show-progress $URL -P models/ -O $FILENAME
+fi
 
-if [ $? -eq 0 ]; then
-  if [ $(md5sum models/$FILENAME | cut -d' ' -f1) = "dd46de7882a7dcbbf46e36f64794680b" ]; then
-    echo -e "${GREEN}Download complete and MD5 hash matches!${NC}"
-  else
-    echo -e "${YELLOW}Download complete but MD5 hash doesn't match. Please try again.${NC}"
-    rm -f models/$FILENAME
-  fi
+if [ $(md5sum models/$FILENAME | cut -d' ' -f1) = "dd46de7882a7dcbbf46e36f64794680b" ]; then
+  echo -e "${GREEN}Download complete and MD5 hash matches!${NC}"
 else
-  echo -e "${YELLOW}Download failed.${NC}"
+  echo -e "${YELLOW}Download complete but MD5 hash doesn't match. Please try again.${NC}"
+  rm -f models/$FILENAME
 fi
